Intermost-Backend/requirements.txt
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0

# MongoDB
pymongo>=4.6.0
dnspython>=2.4.0

# LangChain and AI
langchain>=1.0.0
langchain-core>=1.0.0
langchain-google-genai>=0.0.6
langchain-text-splitters>=1.0.0
langchain-community>=0.3.0
google-generativeai>=0.3.2
numpy>=2.0.0

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# Image handling
Pillow>=10.0.0

# Cloudinary (image hosting)
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

# Utilities
python-dotenv>=1.0.0
whitenoise>=6.6.0
gunicorn>=21.0.0

django-filter>=24.0
drf-yasg
